# storage/api-docs/v2-api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Brillio API",
        "description": "API Backend for Brillio Orientation System",
        "contact": {
            "email": "contact@brillio.africa"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Brillio API Server"
        }
    ],
    "paths": {
...
        "/api/v2/mentors": {
            "get": {
                "tags": [
                    "Mentors"
                ],
                "summary": "Liste des mentors publiés",
                "operationId": "6e859f8dbf633cd8c8f43cd5a47429ef",
                "parameters": [
                    {
                        "name": "specialization",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des mentors"
                    }
                }
            }
        },
        "/api/v2/mentors/{id}": {
            "get": {
                "tags": [
                    "Mentors"
                ],
                "summary": "Détail d'un mentor",
                "operationId": "10001cd02d54e044df378ceca3be06e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail du mentor"
                    },
                    "404": {
                        "description": "Mentor non trouvé"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}